moved ChargerStationSettingsAccordian to its own file
